# docs/agent-platform/agent-platform-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Agent Platform Core API
  version: 1.4.0
  description: Core development contract for Agent, Release, Session, Run, Event, Approval and Artifact.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/agents:
    post:
      tags: [Agents]
      operationId: createAgent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreateRequest'
      responses:
        '201':
          description: Agent created.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ErrorResponse'
    get:
      tags: [Agents]
      operationId: listAgents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: Agent page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPage'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      tags: [Agents]
      operationId: getAgent
      responses:
        '200':
          description: Agent detail.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Agents]
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdateRequest'
      responses:
        '200':
          description: Agent updated.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Agents]
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '202':
          description: Agent deletion workflow accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAccepted'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/copy:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Agents]
      operationId: copyAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyAgentRequest'
      responses:
        '201':
          description: Copied Agent draft.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/disable:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/IfMatch'
    post:
      tags: [Agents]
      operationId: disableAgent
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisableAgentRequest'
      responses:
        '200':
          description: Effective Agent state.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/agents/{agent_id}/references:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      tags: [Agents]
      operationId: listAgentReferences
      responses:
        '200':
          description: References that affect disable or deletion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferencePage'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceAuth:
      type: mutualTLS
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"rv:[1-9][0-9]*"$'
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
...
  headers:
    ETag:
      description: Strong resource version ETag. Its exact value is `"rv:<resource_version>"`.
      schema:
        type: string
        pattern: '^"rv:[1-9][0-9]*"$'
  responses:
    ErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
    AgentCreateRequest:
      type: object
      additionalProperties: false
      required: [code, name, runtime_type]
      properties:
        code:
          type: string
          pattern: '^[a-z][a-z0-9_-]{2,63}$'
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        runtime_type:
          type: string
          enum: [agentscope, codex]
        visibility:
          type: string
          enum: [private, tenant]
          default: private
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 32
    CopyAgentRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code:
          type: string
          pattern: '^[a-z][a-z0-9_-]{2,63}$'
        name:
          type: string
          minLength: 1
          maxLength: 100
    DisableAgentRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 500
    AgentUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        visibility:
          type: string
          enum: [private, tenant]
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 32
        bindings:
          $ref: '#/components/schemas/AgentBindingList'
    ModelRoutingBindingConfiguration:
      type: object
      additionalProperties: false
      required: [fallback_error_codes]
      properties:
        fallback_error_codes:
          type: array
          minItems: 1
          maxItems: 2
          uniqueItems: true
          items:
            type: string
            enum: [RATE_LIMITED, PROVIDER_UNAVAILABLE]
    ResourceBinding:
      type: object
      additionalProperties: false
      required: [resource_type, resource_id, version_policy]
      properties:
        resource_type:
          type: string
          enum: [prompt, skill, mcp, model, knowledge, sandbox, agent]
        resource_id:
          type: string
        version_policy:
          type: string
          enum: [fixed, resolve_on_publish]
        version_id:
          type: [string, 'null']
        binding_role:
          type: [string, 'null']
          enum: [primary, fallback_1, fallback_2, null]
        configuration_schema_version:
          type: [string, 'null']
          enum: [model-routing/v1, null]
        configuration:
          oneOf:
            - $ref: '#/components/schemas/ModelRoutingBindingConfiguration'
            - type: 'null'
      allOf:
        - if:
            required: [resource_type]
            properties:
              resource_type:
                not: { const: model }
          then:
            properties:
              binding_role: { type: 'null' }
              configuration_schema_version: { type: 'null' }
              configuration: { type: 'null' }
        - if:
            required: [binding_role]
            properties:
              binding_role: { type: string }
          then:
            properties:
              resource_type: { const: model }
        - if:
            required: [configuration_schema_version]
            properties:
              configuration_schema_version: { const: model-routing/v1 }
          then:
            required: [binding_role, configuration]
            properties:
              resource_type: { const: model }
              binding_role: { const: primary }
              configuration:
                $ref: '#/components/schemas/ModelRoutingBindingConfiguration'
        - if:
            required: [configuration]
            properties:
              configuration: { type: object }
          then:
            required: [binding_role, configuration_schema_version]
            properties:
              resource_type: { const: model }
              binding_role: { const: primary }
              configuration_schema_version: { const: model-routing/v1 }
    AgentBindingList:
      type: array
      items:
        $ref: '#/components/schemas/ResourceBinding'
      allOf:
        - not:
            contains:
              type: object
              required: [resource_type]
              properties:
                resource_type: { const: model }
            minContains: 4
        - not:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: primary }
            minContains: 2
        - not:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: fallback_1 }
            minContains: 2
        - not:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: fallback_2 }
            minContains: 2
        - if:
            contains:
              type: object
              required: [resource_type]
              properties:
                resource_type: { const: model }
            minContains: 2
          then:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: primary }
            minContains: 1
            maxContains: 1
            items:
              allOf:
                - if:
                    required: [resource_type]
                    properties:
                      resource_type: { const: model }
                  then:
                    required: [binding_role]
                    properties:
                      binding_role:
                        type: string
                        enum: [primary, fallback_1, fallback_2]
        - if:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: fallback_1 }
          then:
            contains:
              type: object
              required:
                - resource_type
                - binding_role
                - configuration_schema_version
                - configuration
              properties:
                resource_type: { const: model }
                binding_role: { const: primary }
                configuration_schema_version: { const: model-routing/v1 }
                configuration:
                  $ref: '#/components/schemas/ModelRoutingBindingConfiguration'
            minContains: 1
            maxContains: 1
        - if:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: fallback_2 }
          then:
            contains:
              type: object
              required: [resource_type, binding_role]
              properties:
                resource_type: { const: model }
                binding_role: { const: fallback_1 }
            minContains: 1
            maxContains: 1
    Agent:
      type: object
      additionalProperties: false
      required: [id, code, name, runtime_type, visibility, tags, bindings, status, resource_version, created_at, updated_at]
      properties:
        id: { type: string }
        code: { type: string, pattern: '^[a-z][a-z0-9_-]{2,63}$' }
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: [string, 'null'], maxLength: 2000 }
        runtime_type: { type: string, enum: [agentscope, codex] }
        visibility: { type: string, enum: [private, tenant] }
        tags:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 32 }
        bindings:
          $ref: '#/components/schemas/AgentBindingList'
        status: { type: string, enum: [DRAFT, ACTIVE, DISABLED, DELETING, DELETED] }
        resource_version: { type: integer, minimum: 1 }
        active_deployment_id: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AgentPage:
      type: object
      additionalProperties: false
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
...
    Reference:
      type: object
      additionalProperties: false
      required: [resource_type, resource_id, reference_type]
      properties:
        resource_type: { type: string }
        resource_id: { type: string }
        reference_type: { enum: [deployment, schedule, child_agent, session, mcp_entry] }
    ReferencePage:
      type: object
      additionalProperties: false
      required: [items, has_more]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Reference' } }
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
...
    OperationAccepted:
      type: object
      additionalProperties: false
      required: [operation_id, status, status_url]
      properties:
        operation_id:
          type: string
        status:
          type: string
          enum: [ACCEPTED, RUNNING]
        status_url:
          type: string
...